# generated/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/meetings": {
      "post": {
        "operationId": "MeetingsController_create",
        "summary": "Создать новую встречу",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMeetingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Встреча успешно создана",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "tags": [
          "meetings"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "MeetingsController_findAll",
        "summary": "Get meetings for the current user",
        "description": "Returns meetings where the user is a participant. Optionally scope to a specific project with ?projectId=.",
        "parameters": [
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filter by meeting status",
            "schema": {
              "enum": [
                "current",
                "past",
                "upcoming"
              ],
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "description": "Scope results to a specific project",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Meeting list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MeetingResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "meetings"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/meetings/{id}": {
      "get": {
        "operationId": "MeetingsController_findOne",
        "summary": "Получить встречу по ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID встречи",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Детали встречи",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          },
          "403": {
            "description": "Нет доступа к встрече"
          },
          "404": {
            "description": "Встреча не найдена"
          }
        },
        "tags": [
          "meetings"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "MeetingsController_update",
        "summary": "Обновить встречу (только создатель)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID встречи",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeetingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Встреча обновлена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          },
          "403": {
            "description": "Только создатель может обновить встречу"
          }
        },
        "tags": [
          "meetings"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "MeetingsController_remove",
        "summary": "Удалить встречу (только создатель)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID встречи",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Встреча удалена"
          },
          "401": {
            "description": "Не авторизован"
          },
          "403": {
            "description": "Только создатель может удалить встречу"
          }
        },
        "tags": [
          "meetings"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/meetings/{id}/phase": {
      "patch": {
        "operationId": "MeetingsController_changePhase",
        "summary": "Изменить фазу встречи (только создатель)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID встречи",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePhaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Фаза встречи изменена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          },
          "403": {
            "description": "Только создатель может менять фазу"
          }
        },
        "tags": [
          "meetings"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/meetings/{id}/join": {
      "post": {
        "operationId": "MeetingsController_joinMeeting",
        "summary": "[DEPRECATED] Присоединиться к встрече (войти в комнату)",
        "deprecated": true,
        "description": "\n      ⚠️ DEPRECATED: Используйте WebSocket event 'join_meeting' вместо этого REST endpoint.\n      \n      Этот endpoint оставлен для обратной совместимости, но не обеспечивает \n      надежное отслеживание присутствия в реальном времени.\n      \n      Для корректной работы голосования используйте Socket.IO:\n      socket.emit('join_meeting', { meetingId })\n    ",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID встречи",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Участник присоединился к встрече (устаревший метод)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          },
          "403": {
            "description": "Только участники могут присоединиться к встрече"
          }
        },
        "tags": [
          "meetings"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/meetings/{id}/leave": {
      "post": {
        "operationId": "MeetingsController_leaveMeeting",
        "summary": "[DEPRECATED] Покинуть встречу (выйти из комнаты)",
        "deprecated": true,
        "description": "\n      ⚠️ DEPRECATED: Используйте WebSocket event 'leave_meeting' вместо этого REST endpoint.\n      \n      Для корректной работы голосования используйте Socket.IO:\n      socket.emit('leave_meeting', { meetingId })\n    ",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID встречи",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Участник покинул встречу (устаревший метод)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "tags": [
          "meetings"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/meetings/{id}/emotional-evaluations": {
      "post": {
        "operationId": "MeetingsController_submitEmotionalEvaluation",
        "summary": "Отправить эмоциональную оценку (все участники включая создателя)",
        "description": "\n      Отправить эмоциональные оценки других участников.\n      \n      ВАЖНО: Голосование полностью опциональное!\n      - Можно отправить пустой массив [] (не голосовать)\n      - Можно отправить оценку только для тех, кого хотите оценить\n      - Можно обновить оценку, отправив новую\n      - Создатель видит все оценки (включая пустые) в /all-submissions\n    ",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID встречи",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitEmotionalEvaluationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Эмоциональная оценка отправлена (может быть пустой)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "tags": [
          "meetings"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/meetings/{id}/understanding-contributions": {
      "post": {
        "operationId": "MeetingsController_submitUnderstandingContribution",
        "summary": "Отправить понимание и вклад (все участники включая создателя)",
        "description": "\n      Отправить самооценку понимания и распределение вклада участников.\n      \n      ВАЖНО: Голосование полностью опциональное!\n      - Можно отправить пустой массив contributions: [] (не голосовать)\n      - Можно отправить вклад только для тех, кого хотите оценить\n      - Проценты не обязаны суммироваться в 100%\n      - Можно обновить оценку, отправив новую\n      - Создатель видит все оценки (включая пустые) в /all-submissions\n    ",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID встречи",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitUnderstandingContributionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Понимание и вклад отправлены (может быть пустым)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "tags": [
          "meetings"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/meetings/{id}/task-plannings": {
      "post": {
        "operationId": "MeetingsController_submitTaskPlanning",
        "summary": "Отправить планирование задачи (фаза task_planning, все участники включая создателя)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID встречи",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitTaskPlanningDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Планирование задачи отправлено и задача создана",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Встреча не в фазе task_planning"
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "tags": [
          "meetings"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/meetings/{id}/task-evaluation-analytics": {
      "get": {
        "operationId": "MeetingsController_getTaskEvaluationAnalytics",
        "summary": "Получить аналитику оценок задач (только создатель)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID встречи",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Аналитика оценок задач с агрегированными данными"
          },
          "401": {
            "description": "Не авторизован"
          },
          "403": {
            "description": "Только создатель может просматривать аналитику"
          }
        },
        "tags": [
          "meetings"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Meetings Quality API",
    "description": "API для платформы отслеживания качества встреч",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateMeetingDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string",
            "description": "Project this meeting belongs to",
            "example": "507f1f77bcf86cd799439050"
          },
          "title": {
            "type": "string",
            "description": "Название встречи",
            "example": "Обсуждение нового проекта"
          },
          "question": {
            "type": "string",
            "description": "Вопрос для обсуждения",
            "example": "Какие технологии использовать для нового проекта?"
          },
          "upcomingDate": {
            "type": "string",
            "description": "Дата и время встречи",
            "example": "2026-02-11T00:00:00.000Z"
          },
          "participantIds": {
            "description": "ID участников встречи",
            "example": [
              "507f1f77bcf86cd799439011",
              "507f1f77bcf86cd799439012"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "question",
          "upcomingDate"
        ]
      },
...
    }
  }
}